fix: rework and clean up the workButtons component and separate logic into smaller reuseable components
